more integers

+++ EEG/ERP.h
@@ -2,7 +2,7 @@
 #define _ERP_h_
 /* ERP.h
  *
- * Copyright (C) 2011,2014,2015 Paul Boersma
+ * Copyright (C) 2011,2014,2015,2017 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,7 +25,7 @@
 /**
 	Look up the channel number from its name.
 */
-long ERP_getChannelNumber (ERP me, const char32 *channelName);
+integer ERP_getChannelNumber (ERP me, const char32 *channelName);
 
 /**
  * Draw the scalp distribution.
@@ -37,7 +37,7 @@ void ERP_drawScalp (ERP me,
 	enum kGraphics_colourScale colourScale, bool garnish);
 void ERP_drawScalp_garnish (Graphics graphics, double vmin, double vmax, enum kGraphics_colourScale colourScale);
 
-void ERP_drawChannel_number (ERP me, Graphics graphics, long channelNumber, double tmin, double tmax, double vmin, double vmax, bool garnish);
+void ERP_drawChannel_number (ERP me, Graphics graphics, integer channelNumber, double tmin, double tmax, double vmin, double vmax, bool garnish);
 void ERP_drawChannel_name (ERP me, Graphics graphics, const char32 *channelName, double tmin, double tmax, double vmin, double vmax, bool garnish);
 
 autoTable ERP_tabulate (ERP me, bool includeSampleNumbers, bool includeTime, int timeDecimals, int voltageDecimals, int units);
